requirements.txt
# Brain Tumor Classification - Requirements
# Optimized for NVIDIA RTX 4050 (4GB VRAM)

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision & Image Processing
timm==1.0.9
opencv-python-headless==4.10.0.84
Pillow>=10.0.0

# Metrics & Evaluation
torchmetrics==1.4.0.post0
scikit-learn==1.5.2

# Utilities
einops==0.8.0
tqdm==4.66.5
numpy>=1.24.0

# Data Download
kaggle>=1.6.0

# Visualization (Optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress & Logging
tensorboard>=2.14.0

# Dashboard
streamlit>=1.30.0

